pyproject.toml
[tool.poetry]
name = "graphorrhea"
version = "0.1.0"
description = ""
authors = ["Nick Beeuwsaert"]
packages = [
    { include = "frontmatter", from = "src" },
    { include = "graphorrhea", from = "src" },
    { include = "plaster_toml", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
pyramid = "^2.0"
pyramid-jinja2 = "^2.8"
pyramid-openapi3 = "^0.13"
pyramid-retry = "^2.1.1"
pyramid-tm = "^2.4"
waitress = "^1.4.4"
acidfs = "^1.0" # TODO: Replace this with custom implementation
docutils = "^0.16"
sqlalchemy = "^1.4.0"
"zope.sqlalchemy" = "^1.3"
passlib = "^1.7.4"
plaster = "^1.0"
toml = "^0.10.2"
colander = "^1.8.3"
"zope.interface" = "^5.3.0"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
# Testing
pytest = "^5.2"
pytest-testmon = "^1.0.3"
pytest-cov = "^2.11.1"
WebTest = "^2.0.35"

# Code style / Linting
black = "^20.8b1"
pydocstyle = "^5.1.1"
flake8 = "^3.8.4"
mypy = "^0.812"

# Debugging
pyramid-debugtoolbar = "^4.9"

# Database management
alembic = "^1.5.7"

# Documentation
Sphinx = "^3.5.2"

[tool.poetry.plugins."plaster.loader_factory"]
"file+toml" = "plaster_toml:Loader"

[tool.poetry.plugins."plaster.wsgi_loader_factory"]
"file+toml" = "plaster_toml:Loader"

[tool.poetry.scripts]
graphorrhea_create_user = "graphorrhea.scripts.create_user:main"

[tool.isort]
known_first_party = [
    "graphorrhea",
    "frontmatter",
    "plaster_toml"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
